19 10 2022 booking

+++ src/main/java/com/ideas2it/bookmymovie/dto/TicketDto.java
@@ -4,14 +4,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
 @Setter
 @NoArgsConstructor
 public class TicketDto {
-    private BigDecimal ticketId;
+    private int ticketId;
     private int noOfSeats;
     private boolean ticketStatus;
     private List<SeatDto> seats;
